pyproject.toml
[tool.poetry]
name = "pyomg"
version = "0.1.0"
description = ""
authors = ["Michael Helmling <michaelhelmling@posteo.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pytaglib = "^2.0.0"
fs = "^2.4.16"
musicbrainzngs = "^0.7.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
